move visa to domain
