+++ src/components/VisaFormSection.tsx
@@ -4,7 +4,7 @@ import {
   getOverallPromptIndex,
   nextStepOfForm,
 } from '@lib/domain/form'
-import { Qualifications, encodeQualifications } from '@lib/visa/visa'
+import { Qualifications, encodeQualifications } from '@lib/domain/visa'
 import { VisaFormPrompt } from '@components/VisaFormPrompt'
 import { useParams, useRouter } from 'next/navigation'
 import { useLanguage } from '@lib/hooks'
