+++ src/lib/domain/errors.ts
